+++ R/output.R
@@ -1,13 +1,5 @@
 
 
-write_stdout <- function(text) {
-  write(text, stdout())
-}
-
-write_stderr <- function(text) {
-  write(text, stderr())
-}
-
 remap_output_streams <- function() {
   output <- import("rpytools.output")
   output$remap_output_streams(