+++ NEWS.md
@@ -30,6 +30,9 @@ Install the development version with: `install_github("rstudio/reticulate")`
 
 - Added `eng_python()` (knitr engine for Python chunks)
 
+- Remove `dim` methods for NumPy arrays (semantics of NumPy reshaping are
+  different from R reshaping)
+
 
 ## reticulate 1.2 (CRAN)