+++ NEWS.md
@@ -45,6 +45,9 @@ Install the development version with: `install_github("rstudio/reticulate")`
 
 - Catch and print errors which occur in generator functions
 
+- Write using Rprintf when providing custom Python output streams
+  (enables correct handling of terminal control characters)
+
 - Implement `isatty` when providing custom Python output streams